+++ App/AutomataExistencias/AutomataExistencias.DataAccess/Aldebaran/ItemByColor.cs
@@ -11,10 +11,10 @@ namespace AutomataExistencias.DataAccess.Aldebaran
         public int Id { get; set; }
 
         [Column("IDITEMXCOLOR")]
-        public short ColorItemId { get; set; }
+        public int ColorItemId { get; set; }
 
         [Column("IDITEM")]
-        public short? ItemId { get; set; }
+        public int? ItemId { get; set; }
 
         [Column("REFITEMXCOLOR")]
         public string ItemByColorReference { get; set; }
